+++ doc/changelog.rst
@@ -5,6 +5,7 @@ Changelog
 **dev**
 
 * When using Voronoi volumes to compute volume of defect cluster, sum volumes of vacancies as well as interstitials/antisites
+* Add option to trace the positions of atoms in a filter list
 
 **v140929**